some renaming and small things
